divided tests in blog posts
