+++ cypress/integration/v2/blocks/blog-posts/desktop-advanced.spec.js
@@ -1,0 +1,6 @@
+/**
+ * External dependencies
+ */
+import { registerTests } from '~stackable-e2e/helpers'
+import { desktopAdvanced } from './index'
+describe( 'Blog Posts Block ( Desktop advanced )', registerTests( [ desktopAdvanced ] ) )
