+++ cypress/integration/v2/blocks/blog-posts/desktop-style.spec.js
@@ -1,0 +1,6 @@
+/**
+ * External dependencies
+ */
+import { registerTests } from '~stackable-e2e/helpers'
+import { desktopStyle } from './index'
+describe( 'Blog Posts Block ( Desktop style )', registerTests( [ desktopStyle ] ) )
