+++ cypress/integration/v2/blocks/blog-posts/mobile-advanced.spec.js
@@ -1,0 +1,6 @@
+/**
+ * External dependencies
+ */
+import { registerTests } from '~stackable-e2e/helpers'
+import { mobileAdvanced } from './index'
+describe( 'Blog Posts Block ( Mobile advanced )', registerTests( [ mobileAdvanced ] ) )
