+++ cypress/integration/v2/blocks/blog-posts/mobile-style.spec.js
@@ -1,0 +1,6 @@
+/**
+ * External dependencies
+ */
+import { registerTests } from '~stackable-e2e/helpers'
+import { mobileStyle } from './index'
+describe( 'Blog Posts Block ( Mobile style )', registerTests( [ mobileStyle ] ) )
