+++ cypress/integration/v2/blocks/blog-posts/tablet-advanced.spec.js
@@ -1,0 +1,6 @@
+/**
+ * External dependencies
+ */
+import { registerTests } from '~stackable-e2e/helpers'
+import { tabletAdvanced } from './index'
+describe( 'Blog Posts Block ( Tablet advanced )', registerTests( [ tabletAdvanced ] ) )
