+++ cypress/integration/v2/blocks/blog-posts/tablet-style.spec.js
@@ -1,0 +1,6 @@
+/**
+ * External dependencies
+ */
+import { registerTests } from '~stackable-e2e/helpers'
+import { tabletStyle } from './index'
+describe( 'Blog Posts Block ( Tablet style )', registerTests( [ tabletStyle ] ) )
