+++ cypress/support/stackable/index.js
@@ -3,4 +3,10 @@ Cypress.on( 'window:before:load', win => {
 	delete win.fetch
 } )
 
+Cypress.on( 'uncaught:exception', () => {
+	// returning false here prevents Cypress from
+	// failing the test
+	return false
+} )
+
 import './commands'
